% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{DrawPage}
\alias{DrawPage}
\title{Draw a single page}
\usage{
DrawPage(subject_id, pg.width, pg.height, before = NULL)
}
\arguments{
\item{pg.width}{- viewport width in pixels}

\item{pg.height}{- viewport height in pixels}

\item{before}{- POSIXt date-time, if not NULL (default), draw
only annotations from this time or earlier.}

\item{subject_index}{index of page to be drawn}
}
\description{
Call from inside a viewport of the size given in the call
}

