% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{FindTypes}
\alias{FindTypes}
\title{Get all the types of annotation}
\usage{
FindTypes(mongo)
}
\arguments{
\item{mongo}{database connection.}
}
\value{
names for each classification type.
}
\description{
Get the names of each type of classification in the database.
}

