% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPageImage}
\alias{GetPageImage}
\title{Get a raster image of a selected page}
\usage{
GetPageImage(subject_id)
}
\arguments{
\item{subject_id}{index of the selected subject (as string)}

\item{mongo}{database connection.}
}
\value{
raster of the image.
}
\description{
Download the image from the website (unless it's already cached).
}

