% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oldWeather4.R
\docType{package}
\name{oldWeather4}
\alias{oldWeather4}
\alias{oldWeather4-package}
\title{oldWeather4 - Process data for oldWeather-whaling}
\description{
Reads the mongodb database created from the project website.
}
\section{functions}{

\itemize{
  \item \code{\link{GetClassifications}} - Extract selected classifications
}
}

\section{data structure}{

}

